Enemies Decrease Timer // Small adjustments
I made some code changes to improve the enemies since they sometimes were bugging and not moving properly. I also made it so that enemies deal dmg (decrease the timer). Also made some material changes. (Also added a samplescene, but that was just for recording some footage for a video)